Modified serve.sh to not try kill ps, commented out broken testcases

+++ testing/serve.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
-pid=$(ps -x | grep 'ng serve' | head -1 | cut -d' ' -f1)
-if [ ! -z "$pid" ]
+num=$(ps -x | grep 'ng serve' -c)
+if [ $num == 2 ]
 then
-	kill $pid
+	pid=$(ps -x | grep 'ng serve' | head -1 | cut -d' ' -f1)
+	if [ ! -z "$pid" ]
+	then
+		kill $pid
+	fi
 fi
 cd ../dashboard
 $(ng config -g cli.warnings.versionMismatch false)
